--- src/components/index.ts
@@ -1,43 +1,0 @@
-import PrismicEmbed from "./PrismicEmbed.vue"
-import PrismicImage from "./PrismicImage.vue"
-import PrismicText from "./PrismicText.vue"
-
-export type { PrismicEmbedProps } from "./PrismicEmbed.vue"
-export type { PrismicImageProps } from "./PrismicImage.vue"
-export type { PrismicTextProps } from "./PrismicText.vue"
-
-export { PrismicEmbed, PrismicImage, PrismicText }
-
-export { usePrismicLink, PrismicLinkImpl, PrismicLink } from "./PrismicLink"
-export type { UsePrismicLinkOptions, PrismicLinkProps } from "./PrismicLink"
-
-export {
-	usePrismicRichText,
-	PrismicRichTextImpl,
-	PrismicRichText,
-} from "./PrismicRichText"
-export type {
-	UsePrismicRichTextOptions,
-	PrismicRichTextProps,
-} from "./PrismicRichText"
-
-export {
-	getSliceComponentProps,
-	TODOSliceComponent,
-	defineSliceZoneComponents,
-	SliceZoneImpl,
-	SliceZone,
-} from "./SliceZone"
-export type {
-	DefineComponentSliceComponentProps,
-	SliceComponentProps,
-	SliceComponentType,
-	SliceLikeRestV2,
-	SliceLikeGraphQL,
-	SliceLike,
-	SliceZoneComponents,
-	SliceZoneResolverArgs,
-	SliceZoneResolver,
-	SliceZoneLike,
-	SliceZoneProps,
-} from "./SliceZone"
